Revert "Revert "Revert "update"""
This reverts commit afadedf37acf8116fccd594dda24958b9c98f47a.

+++ database/migrations/2017_02_17_131502_create_allitems_table.php
@@ -16,7 +16,7 @@ class CreateAllitemsTable extends Migration
         Schema::create('all_items', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('appid');
-            $table->bigInteger('classid');
+            $table->integer('classid');
             $table->integer('instanceid');
             $table->text('icon_url');
             $table->text('icon_url_large');
